requirements.txt
prometheus-client==0.16.0
pyserial==3.5
